Animation Node - MIDI Sync
This node takes a MIDI file in CSV format and creates a set of Empties with F-Curves for the note-on and note-off events from the MIDI file. The node can optionally use the MIDI Velocity, can use an Easing value to slope the F-Curves so note movements are not too sharp and can use an Offset value to move the animation along the Timeline. The Load Sound function also uses this Offset value so the sound and MIDI events are in time.

This release is a Prototype and may contain errors. It is still under development.