MIDI Node To Calculate Indices
This node runs in a separate Node Tree and calculates the match (Index) between Controls Group from my MIDI Write Node and the Keys meshes Group. The new Node tree should be set to No Auto-Execution so this runs only once for the project (Execute by clicking the button). This index list contains a lookup for controls to keys. The Animation Node Tree should be set to run Automatically.

I will do a "Readme" for both nodes when I get some time (Ha Ha) and the code is fully debugged!